requirements.txt
mysql-connector-python==8.0.32
Flask-HTTPAuth
Flask-SQLAlchemy==3.0.3
Flask-Admin
flask==2.2.3
PyMySQL
SQLAlchemy==2.0
Flask-Babel
flask_swagger_ui
Werkzeug==2.2.3
numpy==1.24.3
torch==1.9.0
torchvision==0.10.0
shapely==2.0.1
opencv-python==4.7.0.72
Pillow==9.5.0
humanfriendly==10.0
shapely
PyYAML==6.0
matplotlib==3.7.1
pandas==2.0.2
seaborn==0.12.2
scipy==1.10.1
ultralytics==8.0.145
requests==2.31.0
tqdm==4.65.0
psutil==5.9.5
thop==0.1.1.post2209072238
openpyxl==3.1.2